+++ src/Components/User.php
@@ -31,7 +31,10 @@ class User extends AbstractComponent implements UserInterface
             return $this->data;
         }
 
-        return $this->encode((string) $this->data, self::$subdelimChars);
+        $regexp = '/(?:[^'.self::$unreservedChars.self::$subdelimChars.']+
+            |%(?!'.self::$encodedChars.'))/x';
+
+        return $this->encode((string) $this->data, $regexp);
     }
 
     /**